refactor: 도메인 클래스 패키지 분리
