+++ src/main/java/ladder/domain/LadderGame.java
@@ -1,6 +1,11 @@
 package ladder.domain;
 
 import ladder.domain.creator.LadderCreator;
+import ladder.domain.item.LadderItems;
+import ladder.domain.item.Person;
+import ladder.domain.item.WinningItem;
+import ladder.domain.ladder.Ladder;
+import ladder.domain.ladder.LadderHeight;
 
 import java.util.HashMap;
 import java.util.Map;
