+++ src/main/java/ladder/domain/creator/RandomLadderCreator.java
@@ -1,6 +1,6 @@
 package ladder.domain.creator;
 
-import ladder.domain.Ladder;
+import ladder.domain.ladder.Ladder;
 
 import java.util.stream.IntStream;
