+++ src/test/java/ladder/domain/LadderGameTest.java
@@ -3,6 +3,13 @@ package ladder.domain;
 import ladder.domain.creator.LadderCreator;
 import ladder.domain.creator.RandomLadderCreator;
 import ladder.domain.creator.RandomLineCreator;
+import ladder.domain.item.LadderItems;
+import ladder.domain.item.Person;
+import ladder.domain.item.WinningItem;
+import ladder.domain.ladder.Connection;
+import ladder.domain.ladder.Ladder;
+import ladder.domain.ladder.LadderHeight;
+import ladder.domain.ladder.Line;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
