+++ src/test/java/ladder/domain/item/PersonTest.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.item;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
